Move Android.mk up a level.
The new prebuilt update process will clobber the entire versioned
subdirectory, and it's easier to move this up a level than to add this
file as part of the install.

Also, the way this used to be meant that changing the version number
in the build system would actually pick up the wrong sanitizer
prebuilts.

Change-Id: I3a1fefe861cb3e8cf970594a8c77843b70f22a9a