Update patch set 1
Patch Set 1:

> The revision number seems to have digit too many :)

If you're referring to the part that says "to build 2445339", that's not an svn revision, that's the build number from go/ab. Otherwise I'm not sure what you mean.

 > Why not rebase to r251937?

Rebase hasn't been done yet. This build is just the cherry-picks that we need to fix the windows NDK (and is one that actually came from the build server, so that's something).

Note that I'm not actually moving the platform builds to this atm. That will take some more (read: any) verification.

Patch-set: 1