Update patch set 1
Patch Set 1:

My goal: I have a cmake based build system inside android gerrit where I would like to reference the latest version of clang-tidy in the repository in a way that I don't have to change my code when somebody updates clang (I assume that if a clang is good enough for the platform then it is good enough for my use case as well).

I am not sure if adding a symlink is the right approach but I don't have any other idea so I am open to suggestions. Please let me know what do you think.

(If we agree that this is the right approach then I will create a CL for OSX as well).

Patch-set: 1