Update patch set 8
Patch Set 8:

Setting Sanitize.Never to true (either in the Android.bp, or for less duplication, the go file) would fix that. That's equivalent to what happened in the Android.mk.

Though we should probably be smarter about sanitization and prebuilts -- if it's the same file, the dependencies probably aren't changing base on SANITIZE_TARGET. (Though if we allow specifying a version of the prebuilt with one of the sanitizers turned on, that variant would have different deps)

Patch-set: 8
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>