Export LLVM host libraries in the toolchain
Bug: http://b/64121881

Export libLLVM, libclang, and the libc++ used by them, as prebuilts.
Several annoying/nitty-gritty details need to be handled to export these
transparently:

    - Versions are embedded in all the linux shared libraries, but only
    in the libc++ for darwin.
    - Symlinks with the major version number are created (again
    inconsistently) and are required because the SONAMEs in the actual
    libraries are different from the actual file names, but refer to the
    symlink.
    - Soong does not seem to support Symlink properties for libraries or
    prebuilts.

This export can be cleaned up a bit if the prebuilts include only the
canonicalized libraries (for instance, libclang.so.5 is not a symlink,
but the actual library whose SONAME also states libclang.so.5).  We can
do this packaging for the next full-on clang update.  In the mean time,
this stop-gap solution should unblock moving some host Clang tools to
use the prebuilts rather than outdated sources in external/{llvm,clang}.

Test: Test
https://android-review.googlesource.com/c/platform/development/+/469159
on Linux, Darwin.  Test regular build in aosp/master.

Change-Id: I44c52cb5e816b25152c994e63f4cbeeb2b12c8d2