Update patch set 1
Patch Set 1:

> > Missing device openmp libraries.
 > >
 > > All the asan libraries are static libraries instead of shared,
 > > which the driver is not configured for.
 > 
 > I don't even see some of the libraries where we expect them. It
 > seems like we are indeed not building the sanitizer libraries for
 > many of our architectures (and only building them for the host
 > platform honestly).
 > 
 > >
 > > NDK tests are still building, but that's what I've uncovered so
 > > far.

Ah, this is my bad with packaging. I never bothered to place the actual built runtime libraries into the output directory for stage2. We created everything, but I just missed it. The files are actually in a different out directory, so it should just be a matter of taming things with python.

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>