Update patch set 1
Patch Set 1: -Verified

> > right, and clang++ --version now fails to mention the revision
 > > number, too.
 > >
 > > Android clang version 5.0.1  (based on LLVM 5.0.1svn)
 > 
 > CMake/Upstream-rules fail to find the right revision number because
 > we have downstream patches to build scripts that don't have the SVN
 > tags.  Plus, since we take cherry-picks, it'll end up finding the
 > newest cherry-pick instead of the base revision updated to.
 > 
 > Not sure of a clean way to handle this.  Previously, the base
 > revision was just hard-coded into a pre-generated config.h.

We can change the vendor string, but that might be more annoying as part of the build.py script. This is something we can improve upon later, but it would be best to just get consistent version numbering and then put the relevant info up on a wiki.

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0