Update patch set 1
Patch Set 1:

> This will break scripts using clang-stable symlink to get the
 > current clang prebuild.

It was an emulator boot test that failed. I don't believe that this can possibly be related to this CL. The only official user of the clang-stable symlink is the clang-format repohooks. I would really like to submit this CL, but I will push it for another run through TH first.

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0