Update patch set 1
Patch Set 1: Verified-1

Openmp still doesn't work.

Missing the following patches:

https://android-review.googlesource.com/c/platform/external/openmp_llvm/+/384339
https://android-review.googlesource.com/c/platform/external/openmp_llvm/+/384317

It also now depends on libatomic for arm. Are we sure that got built as arm7 and not arm5?

The ndk-build already links libatomic for you, but it doesn't help in this case because the clang driver places -lomp after the ldflags it was passed. I can hack around this by checking for -fopenmp and manually adding -lomp when we add libgcc/libatomic, but that leaves standalone toolchain and third-party build system users having to deal with this themselves. It's probably long past time to change the clang driver to just always use -latomic anywhere it would do -lgcc.

Patch-set: 1
Reviewer: Gerrit User 1043845 <1043845@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=-1