Update patch set 3
Patch Set 3:

Some __sanitizer_ functions are public interface, even if they are not used in Android right now.

This header, and the others in that directory, declare part of the interface:
https://github.com/llvm-project/llvm-project-20170507/blob/master/compiler-rt/include/sanitizer/common_interface_defs.h

They don't declare everything - there are functions that are used only in the instrumented code.

Patch-set: 3
Reviewer: Gerrit User 1005967 <1005967@85c56323-6fa9-3386-8a01-6480fb634889>