Update patch set 2
Patch Set 2:

It looks a bit excessive. For example, the version in clang-4375569 (upcoming toolchain) does not export _Unwind_*. 
Also these and more:
    __extenddftf2;
    __extendsftf2;

I suggest basing the export list off that or any newer version, it would be a lot smaller. That's when the toolchain switched to upstream cmake build, which correctly applies symbol visibility.

Patch-set: 2