Define static lib modules for ASan/UBSan libraries
This change is the same as c998db3797f0bd73ffabf41b1bc97bdc17e92814
for ASan.

Until now, cc/sanitizer.go have been locating the static libraries
directly by generating the path to the libs, i.e., not via the
libclang_rt_prebuilt_* modules defined in
prebuilts/clang/host/linux-x86/Android.bp). However, with [1], the
dependency to the libraries are now explicitly modeled as
module-to-module dependency. As a result, we need to define modules for
the static variants as well.

Bug: 120894259
Bug: 121038155
Test: m

Change-Id: Idd9ec6ef43a5167bf9902eb8ddd0406656db4ca3