Set sdk_version of the prebuilt libs
When [1] lands, the dependency to these prebuilt libs are correctly
modeled as a module-to-module dependency. Then the dependencies become
subject to the link type check. However, since the prebuilt libs are
without sdk_version, they can't be used from modules built with
sdk_version set. In other worlds, sanitizer build for the modules fail.

Therefore, the prebuilt modules are now with sdk_version. For now, the
version is arbitrarly set to 14 which is sufficiently low, so that the
lib can be used with any module having sdk_version >= 14.

Ideally, the version should be set to the lowest possible API version
number after gathering all external symbol references from each of the
lib.

[1] https://android-review.googlesource.com/c/platform/build/soong/+/853909

Bug: 121358700
Test: m
Change-Id: Ia9e919e72815363e7cc3fa60651dba9e42e0b38b