Add symlink for static ASan/UBSan libraries
The symlink is to give the static library another name since soong does
not allow static/shared variants in prebuilt libraries.

Until now, cc/sanitizer.go have been locating the static libraries
directly by generating the path to the libs, i.e., not via the
libclang_rt_prebuilt_* modules defined in
prebuilts/clang/host/linux-x86/Android.bp). However, with [1], the
dependency to the libraries are now explicitly modeled as
module-to-module dependency. As a result, we need to define modules for
the static variants as well.

[1] https://android-review.googlesource.com/c/platform/build/soong/+/853909

Bug: 120894259
Bug: 121038155
Test: m
Change-Id: I3aa88820c90373102299cee2a844f4b6a7be3a31