Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> When are these used? ASan does not support statically linked executables (which is the reason we have such libraries in hwasan), and otherwise defaults to a shared runtime library on android.

Thanks for pointing this out. Before the change(aosp/853909), even for a static lib, the static runtime library was added to libFlags (sanitize.go:597). However, the libFlags was simply ignored (= not used) when creating an archive for the static lib. So, I would like to argue that the dependency existed though it wasn't respected.

With aosp/853909, the dependency to the static runtime library is made via AddFarVariationDependency(). As a result, if the static variation of the runtime library does not exist (regardless of whether it is actually used or not), the dependency can't be satisfied and the build fails.

I now think that we shouldn't add dependency for static libs except for hwasan. Will fix the build system CL and abandon this CL.

Patch-set: 2
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>