Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> This is still in use by the NDK right now. We need to finish the upgrade to the 'd' build, which I think is blocked behind another bug that Ryan was looking at.

I modified an NDK test to require API 19+ (for consistency with an NDK ASAN test), which should be sufficient to keep NDK tests working when the toolchain is updated. Evgenii fixed the root cause in compiler-rt:

https://reviews.llvm.org/rGc6b5be6cf07858c9a76fbfea3f5a8d9bda1577a2
https://github.com/android-ndk/ndk/issues/999

Maybe we want to cherry-pick that LLVM change for the NDK's next toolchain update? (It's a very small change.)

Patch-set: 1
CC: Gerrit User 1005967 <1005967@85c56323-6fa9-3386-8a01-6480fb634889>