Add libclang_rt.ubsan_standalone as LLNDK.
ubsan_standalone is almost always required for on-device fuzzing. The
problem is, ubsan_standalone isn't available in /system/lib, as it's
only deployed as VNDK.

Add ubsan_standalone as LLNDK so that we can fuzz on-device without issues.

LLNDK stability for ubsan_standalone isn't a problem, as it's not for
production builds (that's ubsan_minimal). See similar change for ASan
here: aosp/498510

This will require the 10.0.1 toolchain, as the mapfiles will be generated
there by aosp/1144750

Test: m -j example_fuzzer && adb sync data && \
adb shell /data/fuzz/example_fuzzer/example_fuzzer

Change-Id: I6b25e0bb1c92ad652db4b8a4e48f5786d69213c7