Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hey Stephen, I'm not sure what your procedure is for adding LLNDK libraries. Can you help me out here?
> 
> (I know you're rolling clang 9.0.8 now, so the mapfiles would have to be updated to the new clang distribution.)

Change http://cs/android/toolchain/llvm_android/build.py?dr=C&g=0&l=474 to generate the map file for ubsan, so it gets created for all future toolchain builds.

We've decided to no longer add/change files in a prebuilt (under clang-r<>) so if this is not time critical, we should hold this change until we roll to 9.0.8 with auto-generated map files.

Patch-set: 1
CC: Gerrit User 1058860 <1058860@85c56323-6fa9-3386-8a01-6480fb634889>