Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > Hey Stephen, I'm not sure what your procedure is for adding LLNDK libraries. Can you help me out here?
> > > 
> > > (I know you're rolling clang 9.0.8 now, so the mapfiles would have to be updated to the new clang distribution.)
> > 
> > Change http://cs/android/toolchain/llvm_android/build.py?dr=C&g=0&l=474 to generate the map file for ubsan, so it gets created for all future toolchain builds.
> > 
> > We've decided to no longer add/change files in a prebuilt (under clang-r<>) so if this is not time critical, we should hold this change until we roll to 9.0.8 with auto-generated map files.
> 
> Pirama said it well. If this isn't time-critical, we'd really prefer to not update any part of the existing prebuilt toolchains. I'm hoping that Nick's update will make it easier to not have to do any manual steps for this anymore.

Happy to wait until the next toolchain pull.

I didn't know that we were auto-generating the new mapfiles! Is that something that's happening for the next toolchain pull? Can you link me to the bug?

My main goal is getting ubsan_standalone as LLNDK in the next toolchain. If I push the LLNDK changes now, it'll fail on the mapfiles. Will there be a mapfile of ubsan_standalone next toolchain, or will this be a chicken-and-egg problem (can't add a mapfile because we have pushed the toolchain, can't set LLNDK because no mapfile)?

Patch-set: 1