Update patch set 2
Patch Set 2:

The NDK's darwin build has been failing consistently with "too many open files" since this was submitted. I have no idea why this would cause it, but it has been consistent since this was submitted.

It's tough to rule out a change in the builder being at fault. Would it be possible to (briefly) revert this to check?

Patch-set: 2
CC: Gerrit User 1043845 <1043845@85c56323-6fa9-3386-8a01-6480fb634889>