Enable no_legacy_features in the cc_toolchain.
This makes the cc_toolchain not use built in legacy features and action
configs, allowing Bazel in Android to create a cc_toolchain from scratch
with the maximum control, e.g. changing libraries_to_link orders.

This CL transliterates the legacy action configs and features for the
parts that the build depends on today.

Bug: 201257475
Test: presubmits
Change-Id: I4bfcd205ccfff043aebf0dfe63c8974ec247e2b3