Convert cpp_std to features.
This creates the cpp_std_default feature, along with one feature each
for the std versions in use for Android. The feature is enabled by
converting the cpp_std attribute value to the feature name in
cc_library_static (which creates the cpp compile action for both
cc_library_shared's root and cc_library itself).

Test: CI
Fixes: 202518741
Change-Id: I2999759d1f3bd67c83bd3f62bedd4e45f21756b4