Revert "Update clang-stable to r468909b."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/77040b56-7bc6-483f-a192-c998201f1b35). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie0c2e90b1b04a407ad9626e96eff27f8fdb8e0ff