Revert "Fix licenses -- dual licensed OR PSF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/d72df58e-0216-4b73-8525-8d9e1206aeb8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia6f297453d892617e1fa66f961b7e955197e5607