Revert "Fix licenses -- dual licensed OR PSF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/c9654139-a1a8-4c26-ad06-c3c4db822466). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I620b89c1cbc071c9965f89ee35d1c1300194f5c9