Revert "Fix licenses -- dual licensed OR PSF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/4bacb34f-761e-473b-b313-fb8ed7adc395). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6508275db0e54a02c40647c11ed5b0ff7df62049