Revert "Fix licenses -- dual licensed OR PSF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/778e52b9-2ea3-46ff-b765-c9365eaa174b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idddd42ff82e614374f3aa2e50e43eaf60880ef78