Revert "make whole archive flags match Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/6aae887b-f36b-490e-b62b-7cd642aa6304). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5fc8772dc34c995ace6cc86939a9b46961d1087d