Revert "Add clang runtimes for riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/3a467718-4e74-4d86-ab29-2ed1882940c5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5e3c50eca2d8a37b8997efff35faaa84f4afe4d6