Revert "Add clang runtimes for riscv64"
Revert "Add clang runtimes for riscv64"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/03a13d16-8257-4504-bfd6-19bcab8b3310). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id8dc0b073caec50aa5a3d90b5262926dea89d393