Revert "Add clang runtimes for riscv64"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/5114e1a2-eb30-4e67-acab-a6f83f1316da). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1371de3785835cab8e8a4ceb616da984f763b5fa