Revert "make whole archive flags match Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/3ad1efc4-b824-4924-a90b-9aa261672981). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I645e7bc4352ed051b8cc2a3253def46c14349583