Revert "Add bpf_compiler_flags feature"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c7e25a05-6fcc-435a-a850-a575c36c2b5b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1d88fc8858408892101ec7e7d2d6ff4637c0ce71