Revert "Remove lib64 -> lib symlink"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c96c3d44-2e86-411d-8695-cc3751b5eaa2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id5023a4af7a401ed81ac528f24d0a7dcd857f44f