Revert "Update prebuilt Clang to r487747 (17.0.0)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f8350a5e-4632-4cc9-9f9e-3e123aa60a88). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia6c2758e977e76f44610e0dcde3942aa5d9cb4b1