Revert "Update prebuilt Clang to r487747 (17.0.0)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/37ae3aa4-1e2c-4e01-a0e4-093c8a1e21da). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I15fc303cde8813a1958ec491801b85934bdd6731