Revert "Update prebuilt Clang to r487747 (17.0.0)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a97a71e4-4a24-4b0a-bcc2-1838e6b15c46). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id71b58d482fbf1fd73007d651e6a748050f17c97