Revert "Update prebuilt Clang to r487747 (17.0.0)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9495362e-7784-4a7b-b20e-1427d311c4fb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3ba716df8abace0aec1b176fbcb0b59f2494192b