Revert "Update prebuilt Clang to r487747 (17.0.0)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1be2b737-86ec-4b97-a221-57e561c9cb5a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5d1460413a4fc4f2ec99c08ab67ce62672d4d26b