Revert "Update prebuilt Clang to r487747 (17.0.0)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8abbe45c-bdbb-41e7-8841-9e36c23644b9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I43c7a29058c42c94da3df41424c0832afe48ffaa