Revert "Update prebuilt Clang to r487747 (17.0.0)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/128f747d-151c-4ee9-8b3b-7c08c3b25b33). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic3688f498b8c03d2c11a312eb596a896b15a5709