Revert "Add r475365b:includes."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7ba6ad0f-8300-41ea-a5a9-ff66106544c2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5788e11590cc92c049a5a2b6edf883f513a429fd