Revert "Add r475365b:includes."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f6a8f657-495c-4162-93d5-829510721a26). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0d90cb1ced08720d2b84b8b95ea0ef86daf2aa8f