Revert "Update prebuilt Clang to r487747 (17.0.0)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4570d72d-34b8-4162-93e9-e9a37340ae32). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I62d4c7a05eb2bd47d4d16b89cf14f58630c2b62a