Revert "Update prebuilt Clang to r487747 (17.0.0)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/64a712b2-c27c-4711-a650-6061336bbb7d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I38fde969151ba573cfb4261e22b6cd64b03d6b74