Revert "Remove dead kleaf symlinks"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/35bd18fd-eb08-4999-97b5-814738de0766). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4eea93f9a4795c928366d8dac684d270e5a92879