Revert "Update README information."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2f127637-16fc-4c19-9a93-5a78fd424d42). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie4b5a37afebb3138fe28c0aa3d7c86f50bfa6e22