Revert "Update README information."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c29d604e-617a-4270-a563-3451ae8a9d6d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9209602734bfc8b311218fa3d79170e70f4fba50