Revert "Remove unused clang-r487747."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f8a04040-1c45-497d-b981-7ea52c74a539). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I28cadb04825a419d643e0935253719deab551446