Revert "Add a toolchain feature for Full LTO"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4265179-7ac0-4b3a-a12b-bb6241cb09f9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib204fbf1f72c833161ef67d2eff45089150aa734