kleaf: SSOT for all supported architectures.
Instead of declaring the architectures in BUILD.bazel,
move it to architecture_constants.bzl so it is easier
to add more toolchains for all supported architectures.

This change intentionally violates DAMP because the
target names are implementation details and should
only be used by register.bzl. Their vsibility is also
set to private.

Bug: 253562056
Test: TH
Change-Id: I4d4aa48c44a36065ebd4c7171050bf96405deac0