Remove unused clang versions
Only keep the Clang versions that are in use for android15+ kernels and
hence remove toolchain versions only used for android14 kernels.

If this change is breaking an android14 kernel, it likely missed out on
branching the tooling to the `main-kernel-build-2023` branch. Apply the
following the change to the manifest to restore this tooling version:

https://r.android.com/q/I5fec1a32869b53459382e94c0839f168d450a97f

Bug: 291278756
Bug: 294190504
Change-Id: Id559cf075c4dfd6d395a6ee0cc05c0993789be36
Signed-off-by: Matthias Maennich <maennich@google.com>