Revert "add prebuilt_build_tool modules for clang tools"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a71275d4-befd-487c-99a4-b131c766c52b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0abec75ad1ab3a66140ebf9e7b373f24e27ef29b