Revert "Update prebuilt Clang to r498229b (17.0.4)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5481324e-8f12-4198-a5f2-db4d56f79f88). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If367be3c3c56b5fdc4bb0139eb7bc0876fb155a0