Revert "symlink 17 -> 18 in clang-r510928/lib/clang/"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/14fb1b14-b331-4c3c-ab25-dcc16b91b593). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6db52a72a8ad0282f6e1cbea2e810f669e1a36e0