Revert "symlink 17 -> 18 in clang-r510928/lib/clang/"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/67deb65f-cb1e-4d8e-ab86-f23785483768). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I64861b674dbb7655594f5cfef3ffc0f1f22ea684