kleaf: Use clang++ to link.
Run clang++ -fuse-ld=lld, instead of
clang -fuse-ld=lld, to link. This is so that
libc++ is added when -stdlib=libc++ is set.

This only has an effect on cc_* rules, but not
on kernel_* rules, because _setup_env.sh sets
LD=ld.lld. It does not infer tool names from toolchain
resolution.

Test: TH
Bug: 324206397
Change-Id: I7ea25e8ce6815a5c869426bf33029027c0d340d4