kleaf: Add -stdlib=libc++ -static-libstdc++ to host toolchain.
This is to ensure host toolchain can link to libc++
to build C++ code.

For compilation, this changes C++ code only; this
has no effect on C / assembly code.

For linking, this affects object files built from
both C / assembly code and C++ code.

This adds --stdlib=libc++ -static-libstdc++ to
HOSTLDFLAGS for Kbuild, but not to HOSTCFLAGS, because
HOSTCFLAGS is calculated
by using C_COMPILE_ACTION_NAME.

Using -static-libstdc++ removes the complication of adding
libc++ to runfiles for cc_binary, adjusting rpath, and
packaging libc++ along with the cc_binary when it is
mentioned in a pkg_* or copy_to_dist_dir rule.

Test: TH
Bug: 324206397
Change-Id: I5893c265c4d83ae4fc261e24d5e7cefceafbebc4