Revert "Run bpfmt on Android.bp"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e4436085-8031-474c-b4be-f074e0715cdb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1f52e499f4cec792bc9e3127ec7419a55d11a0bf