Revert "Switch to the libc++ toolchain prebuilt"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f3360644-d24b-437a-b673-3f7f69a1b776). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia96f25f76ce00bc733d629ac09b23317a8f63d2f