Revert "Switch to the libc++ toolchain prebuilt"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c323cbd-2178-4052-8796-e826cca76e34). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I36bacef80a446080bf3adebec9915c042554cfb1