Revert "Switch to the libc++ toolchain prebuilt"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/065afe0f-7a3d-4c30-9892-2a9dfdc73697). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I70657182fd8d25b58e37ccc16f7be26a02138ce6