Revert "Switch to the libc++ toolchain prebuilt"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/677cae4a-0aee-4233-af8d-98cf9dba85ae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3a14879765abf83d79e2c7a293dc52466607668b