Revert "Switch to the libc++ toolchain prebuilt"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cf17be33-4739-4891-a52d-5180d4fcf97c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I984f7ac7c2d4ecc497133f408c294ff967a74175