Revert "Switch to the libc++ toolchain prebuilt"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/61fa00a8-006d-4c18-bbe5-e0747a5d7886). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I72e54880c209bb2dc4699a3e2610b56c4676374b