Revert "Switch to the libc++ toolchain prebuilt"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/57490344-2420-4f34-975d-8310c0d41ee3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia43ea5d263792909d3065162ca6194b0a11fa200