Revert "Switch to the libc++ toolchain prebuilt"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7b43dcdf-0dd9-4391-bdf0-e1e764c7c00c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4d7b2caaf63ab9ffbb8341ee3c00b09e278830a8