Revert "Switch to the libc++ toolchain prebuilt"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5e9bbffa-f5d4-406c-b8fa-f317ddbe06d4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5fff2ff9ec6f12bd370f0860b4ea60c071369fb6