Revert "Update clang-stable to the updated r522817."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4a138343-4f8a-4a24-b0cb-836354af3df0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If231c7fcd222e12b562a44657c9e3f4659897651
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com